requirements.txt
streamlit
google-generativeai
pypdf
python-dotenv
requests
beautifulsoup4
